requirements.txt
aiofiles
alembic
asyncpg
apscheduler
boto3
elastic_apm
fastapi
fastapi-utilities
fastapi_limiter
Flask
invoke
grpcio
langchain
langchain_aws
langchain_core
langchain_google_vertexai
langchain_mcp_adapters
langchain_openai
langchain_experimental
langgraph
langgraph-checkpoint-mongodb
pandas
redis
gunicorn
pillow
pydantic
pydantic_settings
pymongo
python-dotenv
python_jose
pytest
pytest-asyncio
Requests
sqlalchemy
starlette
tqdm
typesense
typing_extensions
uvicorn
websockets
arize-phoenix-otel
openinference-instrumentation-langchain
psutil
psycopg2